Add tests with too many levels and refactoring

+++ spring-boot-jpa/src/main/java/com/jpa/demo/entity/Book.java
@@ -1,6 +1,9 @@
 package com.jpa.demo.entity;
 
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
 import java.math.BigDecimal;
@@ -17,7 +20,7 @@ public class Book {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", unique = true, nullable = false)
-    private Integer id;
+    private Long id;
 
     @Column(nullable = false)
     private String isbn;
